agents/requirements.txt
# ── Agent Requirements (all agents) ────────────────────────────────────────────
uagents>=0.23.6
uagents-core>=0.4.0
openai>=1.30.0
httpx>=0.27.0
python-dotenv
# YouTube agent extras:
google-auth>=2.29.0
google-auth-httplib2
google-api-python-client>=2.130.0
# LinkedIn agent: no extra deps (uses httpx)
# Content agent extras:
pillow>=10.0.0
requests>=2.31.0
